# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Production backend for UnityWorks Vision AI. Owns the Vision OS perception platform and exposes it through an authorized API.",
    "title": "UnityWorks Vision AI",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/platform/organizations": {
      "get": {
        "description": "Every organization on the platform.\n\nPaginated and searchable from the start rather than when it becomes a\nproblem: this list grows with the customer base, and a console that fetches\nall of them works perfectly until the day it does not.",
        "operationId": "list_organizations_api_v1_platform_organizations_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Organizations Api V1 Platform Organizations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Organizations",
        "tags": [
          "platform"
        ]
      },
      "post": {
        "description": "Create a tenant.\n\nThe id is the slug, not a UUID, and that is deliberate. It becomes the\ntenant half of every camera's runtime identity\n(`app.domain.runtime_identity`), it appears in observation partition\nfilenames, and it is what an engineer reads in a log line at three in the\nmorning. `org-acme:cam-01` says something; a hex id does not.\n\nIt is therefore also validated against that module's charset here, at the\nonly place an organization id is ever minted \u2014 so an id that would break\nthe runtime identity round trip cannot enter the database at all.",
        "operationId": "create_organization_api_v1_platform_organizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Organization Api V1 Platform Organizations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Organization",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/v1/platform/organizations/{organization_id}": {
      "get": {
        "operationId": "get_organization_api_v1_platform_organizations__organization_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Organization Api V1 Platform Organizations  Organization Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Organization",
        "tags": [
          "platform"
        ]
      },
      "patch": {
        "description": "Rename an organization. Not its slug, and not its status.\n\n`slug` is immutable for the same reason a restaurant's is, only more so:\nit is the organization id, it is half of every camera's runtime identity,\nand it names observation partitions on disk. Changing it would orphan\nevery one of them.\n\nStatus has its own route below, because a lifecycle change is a different\nkind of act from a rename and must not be reachable by a client that meant\nto correct a typo.",
        "operationId": "update_organization_api_v1_platform_organizations__organization_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Organization Api V1 Platform Organizations  Organization Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Organization",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/v1/platform/organizations/{organization_id}/enter": {
      "post": {
        "description": "Take a full-reach session inside one organization, and record it.\n\n### This is a widening, and it is meant to look like one\n\nEverything else in this module operates on organizations *as objects* \u2014\ntheir existence, their names, their lifecycle. None of it reads a customer's\ndata, and `PlatformOperator` carries nothing that could. This route is the\nsingle deliberate exception, and the shape of it is the argument:\n\n* it is a **POST**, because it is an act rather than a view;\n* it **writes an audit row before it returns**, filed against the customer,\n  so the record exists in the tenant whose data is about to be read;\n* it mints a **separate token** carrying `act: platform_operator`, so every\n  subsequent request is identifiable as part of this entry rather than\n  indistinguishable from the customer's own staff;\n* the reach it grants is **stated, not resolved** \u2014 see\n  `app.authorization.platform.OPERATOR_ENTRY_PERMISSIONS`. Since 2026-09-22\n  that is every permission: the Platform Admin can do anything inside any\n  organization. A suspended organization still refuses writes.\n\nThe property the previous design had, and which is kept: an operator cannot\nread a tenant's data *through the operator door*. `GET /platform/...` still\nreturns counts and never contents. What has changed is that there is now a\nsecond door, it is locked differently, and going through it is recorded.\n\n### Archived organizations are refused\n\nLogin is refused inside an archived organization, and entry is a login by\nanother name. An operator who needs to look at an archived customer's data\nis describing a restore, which is a decision rather than a click.",
        "operationId": "enter_organization_api_v1_platform_organizations__organization_id__enter_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enter Organization Api V1 Platform Organizations  Organization Id  Enter Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enter Organization",
        "tags": [
          "platform"
        ]
      }
    },
...
  }
}